Calculator now takes half months into account

+++ app/controllers/supply_teachers/home_controller.rb
@@ -32,7 +32,7 @@ module SupplyTeachers
 
       @calculator = FTAToPermCalculator::Calculator.new(
         fixed_term_contract_fee: @previous_step.try(:fixed_term_fee).try(:to_i),
-        current_contract_length: @previous_step.try(:current_contract_length).try(:to_i),
+        current_contract_length: @previous_step.try(:current_contract_length).try(:to_f),
       )
     end
   end
